requirements.txt
pygame
PyOpenGL
Pillow
pywavefront
